+++ pkg/analysis_server/lib/src/handler/legacy/edit_bulk_fixes.dart
@@ -72,6 +72,7 @@ class EditBulkFixes extends LegacyHandler {
       var processor = BulkFixProcessor(
         server.instrumentationService,
         workspace,
+        byteStore: server.byteStore,
         codes: codes,
       );
       if (!updatePubspec) {