+++ pkg/analysis_server/test/services/refactoring/legacy/rename_constructor_test.dart
@@ -170,7 +170,7 @@ void f() {
 class A {
   int field = 0;
 
-  A.newName();
+  new newName();
 }
 class B extends A {
   B() : super.newName() {}
@@ -468,7 +468,7 @@ void f() {
     refactoring.newName = 'newName';
     return assertSuccessfulRefactoring('''
 class A {
-  A.newName();
+  new newName();
 
   int field = 0;
 }
@@ -752,7 +752,7 @@ enum E {
 
   factory E.other() => throw 0;
 
-  const E.newName();
+  const new newName();
 }
 ''');
   }
@@ -780,7 +780,7 @@ enum E {
 
   final int foo = 0;
 
-  const E.newName();
+  const new newName();
 }
 ''');
   }
@@ -806,7 +806,7 @@ enum E {
 enum E {
   v1.newName(), v2.newName(), v3.newName();
 
-  const E.newName();
+  const new newName();
 
   void foo() {}
 }
@@ -832,7 +832,7 @@ enum E {
 enum E {
   v1.newName(), v2.newName(), v3.newName();
 
-  const E.newName();
+  const new newName();
 }
 ''');
   }
@@ -856,7 +856,7 @@ enum E {
 enum E {
   v1.newName(), v2.newName(), v3.newName();
 
-  const E.newName();
+  const new newName();
 }
 ''');
   }