+++ runtime/vm/raw_object.h
@@ -2260,9 +2260,7 @@ class UntaggedPcDescriptors : public UntaggedObject {
   V(OsrEntry, kRuntimeCall << 1)                                               \
   /* Call rewind target address. */                                            \
   V(Rewind, kOsrEntry << 1)                                                    \
-  /* Target-word-size relocation. */                                           \
-  V(BSSRelocation, kRewind << 1)                                               \
-  V(Other, kBSSRelocation << 1)                                                \
+  V(Other, kRewind << 1)                                                       \
   V(AnyKind, -1)
 
   enum Kind {